Sync to the original

+++ packages/Webkul/Checkout/src/Cart.php
@@ -185,9 +185,7 @@ class Cart {
                 $result = $this->createItem($id, $data);
             }
 
-            session()->flash('success', trans('shop::checkout.cart.success'));
-
-            return true;
+            return $result;
         } else {
             return $this->create($id, $data);
         }
@@ -325,7 +323,9 @@ class Cart {
 
     /**
      * Update the cartItem on cart checkout page and if already added item is added again
-     *
+     * @param $id product_id of cartItem instance
+     * @param $data new requested quantities by customer
+     * @param $itemId is id from cartItem instance
      * @return boolean
      */
     public function updateItem($id, $data, $itemId)
@@ -371,7 +371,6 @@ class Cart {
 
             return false;
         }
-
     }
 
     /**
@@ -388,7 +387,10 @@ class Cart {
             if($cart->items()->get()->count() == 0) {
                 $this->cart->delete($cart->id);
 
-                $this->deActivateCart();
+                // $this->deActivateCart();
+                if(session()->has('cart')) {
+                    session()->forget('cart');
+                }
             }
 
             session()->flash('success', trans('shop::app.checkout.cart.item.success-remove'));
@@ -692,10 +694,14 @@ class Cart {
      */
     public function collectTotals()
     {
+        $validated = $this->validateItems();
+
+        if(!$validated) {
+            return false;
+        }
+
         if(!$cart = $this->getCart())
             return false;
-
-        $this->validateItems();
 
         $this->calculateItemsTax();
 
@@ -741,11 +747,15 @@ class Cart {
     {
         $cart = $this->getCart();
 
+        if(!$cart) {
+            return false;
+        }
+
         //rare case of accident-->used when there are no items.
         if(count($cart->items) == 0) {
             $this->cart->delete($cart->id);
 
-            return redirect()->route('shop.home.index');
+            return false;
         } else {
             $items = $cart->items;
 
@@ -1078,6 +1088,10 @@ class Cart {
 
                 return $result;
             } else {
+                $data['product'] = $id;
+                $data['is_configurable'] = false;
+                $data['quantity'] = 1;
+
                 $result = $this->add($id, $data);
 
                 return $result;
